+++ packages/reporter/cypress/integration/unit/app_state_spec.ts
@@ -141,28 +141,6 @@ describe('app state', () => {
     })
   })
 
-  context('#setForcingGc', () => {
-    it('sets forcingGc', () => {
-      const instance = new AppState()
-
-      instance.setForcingGc(false)
-      expect(instance.forcingGc).to.be.false
-      instance.setForcingGc(true)
-      expect(instance.forcingGc).to.be.true
-    })
-  })
-
-  context('#setFirefoxGcInterval', () => {
-    it('sets forcingGc', () => {
-      const instance = new AppState()
-
-      instance.setFirefoxGcInterval(111)
-      expect(instance.firefoxGcInterval).to.eq(111)
-      instance.setFirefoxGcInterval(222)
-      expect(instance.firefoxGcInterval).to.eq(222)
-    })
-  })
-
   context('#setStudioActive', () => {
     it('sets studioActive', () => {
       const instance = new AppState()