+++ packages/server/test/unit/modes/run_spec.js
@@ -24,7 +24,7 @@ const ProjectStatic = require(`${root}../lib/project_static`)
 
 describe('lib/modes/run', () => {
   beforeEach(function () {
-    this.projectInstance = new ProjectBase({ projectRoot: '/_test-output/path/to/project-e2e', projectType: 'e2e' })
+    this.projectInstance = new ProjectBase({ projectRoot: '/_test-output/path/to/project-e2e', testingType: 'e2e' })
   })
 
   context('.getProjectId', () => {
@@ -100,9 +100,9 @@ describe('lib/modes/run', () => {
     it('sets width and height', () => {
       const props = runMode.getElectronProps()
 
-      expect(props.width).to.eq(1920)
+      expect(props.width).to.eq(1280)
 
-      expect(props.height).to.eq(1080)
+      expect(props.height).to.eq(720)
     })
 
     it('sets show to boolean', () => {
@@ -699,6 +699,15 @@ describe('lib/modes/run', () => {
 
       return expect(runMode.run({ browser: 'opera' }))
       .to.be.rejectedWith(/invalid browser family in/)
+    })
+
+    it('throws an error if unsupportedVersion', () => {
+      const browser = { displayName: 'SomeBrowser', warning: 'blah blah', unsupportedVersion: true }
+
+      sinon.stub(browsers, 'ensureAndGetByNameOrPath').resolves(browser)
+
+      return expect(runMode.run())
+      .to.be.rejectedWith('blah blah')
     })
 
     it('shows no warnings for chrome browser', () => {